Fixing some issues highlighted by PVS-Studio static code analyzer.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@384 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/UI/Core/ListBox.cpp
@@ -77,7 +77,7 @@ void ListBox::_updateItemDisplay()
 {
 	mItemWidth = rect().width();
 
-	if ((mLineHeight * mItems.size()) > height())
+	if ((mLineHeight * mItems.size()) > static_cast<int>(height()))
 	{
 		mLineCount = static_cast<int>(height() / mLineHeight);
 		if (mLineCount < static_cast<int>(mItems.size()))
